requirements.txt
streamlit
folium
streamlit-folium
plotly
pandas
matplotlib
Pillow
requests